Set auth in clients before they are used
When credentials are preprovisioned, and thus read from a file,
only the fields configured in accounts.yaml are available in the
credentials object. Several tests rely on used_id / project_id
being available in the credentials object. To ensure a consistent
behaviour between dynamic and preprovisioned credentials, setting
auth in the client manager before it is returned to the test, so
that all fields are available.

test_security_groups_basic_ops has its own solution for filling in
IDs, which is broken, so removing it and let it rely on the IDs
being there.

Change-Id: I3450c055c63782a97413606f90ad77b7dba452f1
Partial-bug: #1609156