Avoid volume limit usage for existing volumes
When using pre-provisioned credentials, the tenant may have volumes
other than those created in the testcase and so will make
test_get_volume_absolute_limits fail unexpectedly. e.g. the existing
volumes may be those not being deleted successfully and ending up with
status 'ERROR', which is not rarely seen in production clouds. So this
is to avoid existing volumes in AbsoluteLimitsV2Tests.

Change-Id: Icf2673c2be26be1ff279e019d9cb13ec611ec741