Shared images test create with default visibility
As discussed on the openstack-dev mailing list[1] the
shared image test relies on known bugs[2] in the images API
by creating an image with private visibility and attempting
to share that image while it has private visibility.

This test was recently added without review or consideration
for the known bugs.[3]

The bug fix approved by the community is to introduce a new
visibility status of 'shared' and to use that visibility by
default on new image creation[4]. Therefore in this test,
an image must be created with default visibility to use the
image sharing feature.

Minor cleanup to the test is included to eliminate unnecessary
details and modifications for the tested scenario.

Note: This change is blocking to the implementation[5] of the
spec noted. That spec is blocking other work so this review
is a high priority for the Glance community.

[1] http://lists.openstack.org/pipermail/openstack-dev/2016-December/109370.html
[1] http://lists.openstack.org/pipermail/openstack-dev/2017-January/109678.html
[2] Related-Bug: 1452443
[2] Partial-Bug: 1394299
[3] Ia1369fa73f56ec290e868d3f2f2873785a70951d
[4] http://specs.openstack.org/openstack/glance-specs/specs/newton/approved/glance/community_visibility.html
[5] I94bc7708b291ce37319539e27b3e88c9a17e1a9f
Partially-Implements: bp community-level-v2-image-sharing

Change-Id: I6e3268f3712cbc0aadb51d204c694023b92d55a5