Add a test for reproducing bug/1651064
There is a complex API parameter "block_device_mapping_v2" on Nova API,
the parameter allows multiple combination of attributes and it consists
of multiple dicts.
There was a bug that Nova doesn't validate 2nd dict on the parameter
and an internal error happened if clients pass invalid parameter.
This patch adds a test for reproducing this bug.

Change-Id: I3d4ab22fbf11c38d6b161abd7078a89350c37305
Related-Bug: 1651064
Depends-On: I56348dc2b7abd9526dfbf21305755505765811c5