Modify comments in xxx_name_length_exceeds_256
The testcase name test_create_server_name_length_exceeds_256 and
test_update_server_name_length_exceeds_256 are not correct, because
the max name length is 255, not 256.

Because now we are unable to modify the testnames in defcore because
alias is not supported in older json files,
https://github.com/openstack/defcore/blob/16668dddea39725088040fdef5404a656b24a606/2015.03.json,
so here we just modify the wrong comment to make the code less
confusing.

Change-Id: I67d5227f308e957e7eabd6230a6b9df17540eb5d