Implied roles API tests
Implied roles is a feature added recently to keystone. It enables the
possibility to have roles implying another roles, this avoids the need
to add specific role assignments to each action a user can take. For
example: if we had the "admin" role implying the "member" role, we
would not need to add entries specifiying the "member" role for a
user that is already "admin".

Change-Id: I81b3fce613c80aad9e80847df5a61ed10887221c