Domain specific roles API tests
This patch adds the API tests for the domain specific roles feature.

This feature provides the power to have more granularity in defining the
roles (and implied roles rules) within the same cloud provider, this is
accomplished by allowing the roles to be handled (scoped) in the domain
level. For the "implied roles" use case, a domain specific role can be
used in the same way as a "global" role - the difference is that we
can't have a global role implying a domain specific one.

Another limitation is that domain specific roles can't be used to scope
a token - they don't have an effective power when used in an assignment.

Change-Id: Ic3676973b52630cca2c7edd396bcba828f21d08f