Add resources_prefix to rand_name when created
Tests that use rand_name from tempest.lib.common.data_utils
are not adding the resources_prefix from the tempest
config file when creating resources.  These tests should
be using tempest.common.data_utils which passes in the
resources_prefix as a parameter in the rand_name function call.

When resources_prefix is set in the tempest config file all of
the resources created should begin with this attribute. For some
tests this attribute is not passed to the rand_name function call.
This addresses that issue.

Change-Id: I662aad1916949750581e793981988edf1560ea94
Closes-Bug: #1628016