Add test for compute API microversion 2.42
This just runs the 2.32 test against the 2.42 microversion.
The content is the same. The issue with tags is due to a bug
where block device tags only work for the 2.32 microversion
specifically, and network tags only worked for 2.32->2.36
microversions. Network tags were broken in the 2.37 microversion.
The 2.42 microversion re-introduces block device and network
tags when creating a server.

Depends-On: Ia0869dc6f7f5bd347ccbd0930d1d668d37695a22
Related-Bug: #1658571
Part of blueprint fix-tag-attribute-disappearing

Change-Id: Ide579744f4b2c768f622f1f3d928e9e2b02f1051