Add related bug#1659811 for tenant_id filter compute tests
Bug#1659811 on compute side when admin try to filter servers
with tenant_id but no all_tenant.

On Py3 it did not work due to hash randomization is enabled
by default on Py3.

This commit adds that bug as related bug in respective tests
for tracing the same in future.

Change-Id: I2ae66f257ef52972ff1143c92b30aa8acb27d4ab