Add DeleteErrorException
In function “wait_for_server_termination”，it is in the waiting
for server to reach termination. So it's more reasonable to throw
a deleting error exception here when server going error.
This patch adds a DeleteErrorException and let function
“wait_for_server_termination” to use it. This makes the exception
message more clear when server(or other resources) entering error
status in the waiting for deletion.

And the DeleteErrorException will be used
in change I6f92f00b8fe4651c27ff4da783b004d29c526dfe

Change-Id: I6646ce732a4f3802338c8f2a478fd81ff5a16a55