Nova: test live migration with serial console
Nova offers a "serial console" as an alternative to graphical consoles
like VNC/SPICE/RDP. This is useful for platforms which don't have
graphical consoles, for example the "IBM system z" platform.

This change introduces a test which ensures that the interaction with
the serial console is possible before and after a live-migration.
As the unified remote console API is available since microversion 2.6,
I use this as a base for the tests. This made id necessary to update
the schemas.

This change introduces a config option to enable new test cases for
the serial console.
A Nova change (see I7af395a867e0657c26fa064d2b0134345cd96814),
which uses the hook for live-migration testing, will use the config
option of this change to alter the testing system on the fly to
enable the testing of the serial console.

Closes-Bug: #1560358
Needed-By: I7af395a867e0657c26fa064d2b0134345cd96814
Change-Id: I020fd94d970ad0cdf7ab65d7656da6ca5766094b