Network scenarios: remove some instance variables
In the patch, I removed some instance variables in some network
scenarios: self.security_group, self.network_v6, self.network.

We still have a lot of instance variables, which is a major reason
why reviewing network scenarios is an absolute pain. But this
is a small step in the right direction.

Change-Id: I482a06a6592bf1b402d1faf5e78c1be17f05b6b1