Correct test_list_servers_filtered_by_ip for bug 1668828"
test_list_servers_filtered_by_ip will fail occasionally, because
now on Nova side filter by ip is always "regexp match",
https://github.com/openstack/nova/blob/master/nova/api/openstack/compute/schemas/servers.py#L355
so if server1's ip is "10.1.1.1" and servers2's ip is "10.1.1.10",
filter by server1's ip "10.1.1.1" will get both servers in the list.

This is to use the longest ip for the filter, i.e., if server1's
ip is "10.1.1.1" and server2's ip is "10.1.1.10", then we should use
"10.1.1.10" for the filter, so to ensure only one server is returned
in the list.

Change-Id: I87c325cb80a95861287c54fbd1b5718cfb9ef310
Closes-Bug: #1668828