Add network_mask_bits argument to assign_static_ip()
assign_static_ip() is used in test_network_basic_ops only and it is
not necessary to have a config value of project_network_mask_bits in
RemoteClient.
This patch adds network_mask_bits argument to assign_static_ip() with
the same default value as the config, and it removes the configuration
dependency from RemoteClient()

This is a step for making remote_client stable interface(tempest.lib)
and we need to remove CONF dependency from such module as possible
before defining stable interface.

NOTE: openstack/congress only needs to follow this change.

Change-Id: I430424cc776cb12906c7890cf584797ebea1cd41