Use tempest.lib data_utils - image
Tempest has duplicated modules of data_utils, one is tempest.lib and
the other is tempest.common. The difference is tempest.common module
refers to the config option resources_prefix. And the option is marked
as deprecated already. So it is nice to replace the calls with tempest.lib
one.

Change-Id: I2ae4ca7e132a961ffcb072988551c3c9e7cf8266