Use tempest.lib data_utils - network
Tempest has duplicated modules of data_utils, one is tempest.lib and
the other is tempest.common. The difference is tempest.common module
refers to the config option resources_prefix. And the option is marked
as deprecated already. So it is nice to replace the calls with tempest.lib
one.

Change-Id: I6d8d9fcea9f31a4ce6b3c893383fce12f4d17c18