Use tempest.lib data_utils - object_storage
Tempest has duplicated modules of data_utils, one is tempest.lib and
the other is tempest.common. The difference is tempest.common module
refers to the config option resources_prefix. And the option is marked
as deprecated already. So it is nice to replace the calls with tempest.lib
one.

Change-Id: Ifb59b89ec1bd7f8c17a55026b32b8af35c945594