Use tempest.lib data_utils - volume
Tempest has duplicated modules of data_utils, one is tempest.lib and
the other is tempest.common. The difference is tempest.common module
refers to the config option resources_prefix. And the option is marked
as deprecated already. So it is nice to replace the calls with tempest.lib
one.

Change-Id: I90cd12ac7a8050125956654feb0fa8a221e7958e