Fix race in test_attach_detach_volume
The test boots a server, attaches a volume and then power cycles
the server, to check that the volume is still attached.
When ssh validation is on, we must wait for the ssh key to be
injected in the guest authorized keys file, before power cycling,
otherwise ssh will fail later in the test.

Change-Id: I506085b4efd2501e41e91530b3fedcfad286ef0e