Raise exception when get_disks() cannot get 'TYPE' column
This commit makes get_disks() raise an exception when it cannot get
'TYPE' column. In most of new linux distros, this cannot be happened,
however, in old distros, the 'TYPE' column doesn't exist. In that
situation, this could be useful for debugging and it's easy to
understand what is the root cause.

Change-Id: I328f1c2ea80478a00432f51e77ffd12036829430
Related-Bug: #1673607