Add additional roles method to v3 roles client.
Adds the following API call to the roles_client library:
  - list all role inference rules (/v3/role_inferences)

This enables querying these APIs for a list of all the
role assignments or all the role inferences.

Note that listing all role inference rules is distinct
from listing inference rules. This can be confirmed
in the API docs [0] and in the code itself [1]
(see list_role_inference_rules vs list_implied_roles).

[0] https://developer.openstack.org/api-ref/identity/v3/#list-all-role-inference-rules
[1] https://github.com/openstack/keystone/blob/3e5ead0a45f698eed4162787b723090cee4733f8/keystone/assignment/routers.py

Change-Id: I9d6c0dc83a85bbca173b753183d838adfee04008