Merge Cinder v3 base test module into general one
Cinder v3 base test module contained some v3 specific test classes
but that is not so different from the general one which is used for
Cinder v2 tests.
We will use the general test module for Cinder v3 tests also in the
future. So this patch merges Cinder v3 base test module for re-using
the v2 tests.

Change-Id: Ia956908dff14f15e55687f3daa4789dde096ba41