Deprecate client_parameters from ServiceClients
client_parameters was been designed as a way to pass parameters
for service clients to the ServiceClients class.

After the initial implementation, the client registry was added,
which allows for automatic registration of all clients and their
parameters. If configuration is available, parameters will be
pushed into the registry automatically.

When the registry part was implemented, the client_parameters logic
has been broken. Parameters are loaded and prepared correctly, but
ultimately they are not used for registration, they are basically ignored.

So this commit simply deprecates client_parameters, so that once it's not
used in tempest/clients.py anymore and the deprecation periodic is over
it can be removed.

closes-bug: #1680915
Change-Id: Ib37bde098d15ce440297728f0d0e1087b8693b6c