test_l3_agent_scheduler: remove workaround code for Liberty
Now that an L3 Agent always returns its "agent_mode" no need to fallback
to default value if the agent_mode is not returned.

This fixes an old TODO by armax.

Also, optimize a bit the test to use server-side L3 Agent filtering.
And simplify multistring docstring to comply with H404.

Change-Id: Ib02d04082184bc7d5a520180d3d047d6b2024344