Remove the heat tests
This commit removes the heat tests from tempest, we've been pushing for
these to become a plugin for a long time without any progress. In
addition the heat team has no desire to maintain these tests. [1] Since
they're not providing value anymore let's just delete them. Note, the
clients are not removed in this patch because other projects depend on
those and we shouldn't break those users. Ideally those interfaces
should be moved to lib and just left there.

[1] http://lists.openstack.org/pipermail/openstack-dev/2017-May/116170.html

Depends-On: I983de842f1bf84a939cbcda6c65185c694a69949
Change-Id: Ifc2cac575919da4e361eaa3234d7e0f3e2f98d21