Add max_microversion for compute volumes_extensions_client tests
The ``volumes_extensions_client`` [0] uses a deprecated nova
API, which starts throwing a 404 from microversion 2.36+ [1].

This patch adds max_microversion = '2.35' to compute tests that use the
``volumes_extensions_client`` and the ``snapshots_extensions_client``.

This can be tested by running:

curl -g -i -X GET http://127.0.0.1:8774/v2.1/os-volumes \
-H "OpenStack-API-Version: compute 2.35" \
-H "User-Agent: python-novaclient" \
-H "Accept: application/json" \
-H "X-OpenStack-Nova-API-Version: 2.35" \
-H "X-Auth-Token: $TOKEN"

which passes -- whereas replacing "2.35" with "2.36" above throws
a 404, in accordance with the documentation [1].

This patch also replaces ``volumes_extensions_client`` from
the base compute delete_volume classmethod with the ``volumes_client``.
This is because the functionality is identical: the former client
simply does a proxy call to Cinder, so we should use the non-deprecated
``volumes_client`` instead.

[0] https://github.com/openstack/tempest/blob/master/tempest/clients.py#L174
[1] https://developer.openstack.org/api-ref/compute/#volume-extension-os-volumes-os-snapshots-deprecated

Change-Id: I4a1737aa4456580d8c8cb2df8d1c28792f774104