Use Neutron (if available) to create floating IPs for validation
The 2.36 microversion in the compute API deprecates the os-floating-ips
API since it's a proxy when using Neutron, and nova-network is deprecated.

When ssh validation is enabled in a Tempest run, we should create the
floating IP using Neutron if available and only fallback to the compute
API if using nova-network.

Change-Id: Ib345abdeec17fef85f99beb0ff4f3639e95acf6f