Update server nic handling for LXD style nic names
Nova-LXD/LXD supports hotplugging of network interfaces, but
nic names are a little different to those in qemu/kvm:

  1249: eth0@if1250: <BROADCAST...> mtu 1450 ... 1000

These are actually part of a veth pair (hence the @ annotation).

Update server nic handling to deal with trailing '@if<peerid>'
so that nics get correctly detected when testing LXD based
OpenStack Clouds.

Change-Id: Ic7eb222d627513b9251772e2a3bf4909c4794a9f