Use urllib2 instead of requests in tempest generate plugin list
While using generate-tempest-plugins-list.py in tempest-plugin-sanity
CI jobs fails to run as it requires requests module to parse the git
url and is not install by default in CI job. If we are going to
install requests we need to use sudo for that which is not a
good idea, so it is better to move the script to urllib2.

Change-Id: Ifed51c67c03f993a4833413826ac15104db152fc