Wait for server deletion if validation setup fails
Previously, if validation setup failed, the server would get deleted
and the test would immediately proceed with cleanup. This is a problem
if the server was booted with volumes. The volumes would be deleted
without waiting for the server to go away first, and this would fail
since they are still attached. This patch waits for server termination
before proceeding with cleanup.

Change-Id: I57a9c85de30855091e5443692436034a51886e63