Add network tags-ext API tests
Adds network API tests for the tags-ext network extension. These
tests validates creating, listing, deleting, checking the existence
of, deleting all, and updating all tags for the following resources:
subnets, ports, routers and subnetpools.

The tests have been designed such that adding a new entry to
`SUPPORTED_RESOURCES` at the top of TagsExtTest and creating
the new resource in resource_setup should work -- resulting
in lower code maintenance.

Note that API tests for the networks resource has already been added
in this commit Icfff444ee7638a3220d228330f9162044673636c -- which is
part of the tag extension. This commit uses a separate class which
checks for the tag-ext extension, which all the other aforementioned
resources rely on.

Change-Id: I0df5aa7e3177f70a75fd76d811bed8553fb26c25