Test size extend for an attached volume
This tests a new Cinder microversion to allow extending
the size of an attached volume.

Once the volume is extended in the storage backend, the
Cinder volume manager will call a new microversion in
Nova to let Nova complete the operation for the server
instance that is attached to the volume.

The test initiates the operation through the Cinder API
and then monitors the Nova API until completion.

Related to cinder blueprint extend-attached-volume
Related to nova blueprint nova-support-attached-volume-extend

Change-Id: Ibace6c2f91be9753a44e5f79fd013df11654851b
Depends-On: I60c8ea9eb0bbcfe41f5f0a30ed8dc67bdcab3ebc