Fix instance_action_events response schema for finish_time and result
An instance action event's finish_time and result can be None when
the event is not yet completed. This is common in Nova. The nova-api
service creates the instance action record and the nova-compute service
has a method which performs the actual operation on the server
instance. The method in the compute is wrapped with an EventReporter
decorator which records the event start and finish times. So until
the event (operation) is complete, the event finish_time and result
are  going to be None.

Change-Id: I964d9170ca7f038e9ab968e119f03fbe9efae588
Closes-Bug: #1702758