Add flag to generate tempest plugin list
This commit adds a flag to generate a tempest plugin list. In a local
environment, we don't need to generate a tempest plugin list, always
when we just need to see the other documents. Besides, generating the
tempest plugin list takes a long time. When GENERATE_TEMPEST_PLUGIN_LIST
environment variable is 'true'(default), the plugin list is created.
When the other value such as 'false' is specified, it isn't.

Change-Id: I6fd3cb3045ba8bbfb39c597413e13faa3cd24ae3