Move multi-nic tests to dedicated module
The test_create_server module creates a server and then it runs
multiple checks in different tests. This is done for DiskConfig
AUTO and MANUAL.

The two multinic tests however create their own server, so they
should not belong to the same class. With the current setup we
run them twice, once per DiskConfig value, which uses gate time
and resources for no clear purpose.

To fix this I'm moving these two multi nic tests to a dedicated
module.

Change-Id: I97dd4b5c703d41d9084e6e049144e4df4e3fb75d