Cleanup the orchestration client init and attr
Clients.py inits the orchestration client and defines and attribute
in the client manager for it. The only known consumers of attribute
are Tempest itself in the cleanup service and Murano.

I proposed a patch to Murano: Ic6fef4b4f260d13efd7b035bc012369f0da55602

The aim is to make clients stable for plugins, and I don't want the
Heat client to be in the stable interface since it's not something
that we will be supporting on Tempest side: all the Heat config
values are deprecated and will be removed during Queens.

Depends-on: Ic6fef4b4f260d13efd7b035bc012369f0da55602
Change-Id: I9f4703296a44e741cab8db6b52ba410b7dd3bed6