Remove creds providers dependency from clients
We want to make cred providers a stable interface and move it to
lib but to do so it must not depend on modules outside of lib or
not generally stable otherwise. The last dependency left is
tempest.clients, which can be replaced by its tempest.lib
counterpart, tempest.lib.services.clients.

While clients does not depend on configuration directly, it uses
the client registry singleton. When this class is used as part of
tempest, it's populated with configuration values the first time
the CONF object is instantiated, so that Tempest configuration
values are still honoured.

When DynamicCredentialProvider is used by a consumer that does
not have a Tempest config file, all the settings defined in
__init__ will be honoured. To control more settings, the
client registry must be populated before the Credentials
Provider is used with the appropriate settings.

Change-Id: I7b262607a1fa9f67e6b1d3ec2c4cf1ccffb952a6