Remove specific project checking in test
In test_create_is_domain_project, the test would check
for a specific project object within a list of returned
projects. If a new resource attribute is added to
projects, such as "tags", this test will fail due to
the sample project not containing the "tags" atttibute.

If we simply add a "tags" attribute to the sample project
generation, then the test will still fail for anything before
"tags" is added to keystone. After tags is added, this test
would also still potentially fail due to how tags is treated
as a list of strings, which is not guarenteed to be in order.

If two instances of the same project contained the same tags,
but not in the same order, they would be considered not equal
when compared. This would cause this test to fail randomly.

This change removes the checking of a specific project in a
list of projects to checking if the project ids match similar
to how domains are compared below this check. This prevents this
test from failing whenever an attribute of project is changed or
if the attribute is a list with no guarenteed order.

Change-Id: I00f094a5584be40ab477cbf680a5f6d1afb4d21b
Partially-Implements: bp project-tags