Increase unit test coverage for v2 images client
This patch adds unit tests for the v2 images client. Note that
``test_show_image_file`` currently expects a response body
of {} but this is because it is not possible to use
data_utils.rand_bytes() since ``create_response``
will always try to then execute json.dumps which will result
in a JSON-related error being thrown. (A follow up patch
would be needed to fix this issue.)

Partially Implements: blueprint tempest-lib-missing-test-coverage

Change-Id: I33a718d8ecf133a55c76dccb6746232ee6152ca2