Add compute test for ServerGroupAntiAffinityFilter
We have a test for the ServerGroupAffinityFilter but
we don't have one for ServerGroupAntiAffinityFilter
which requires two compute hosts. As the FilterScheduler
in Nova relies more and more on the Placement
service and allocation tracking between multiple compute
nodes, and we have started blocking up-calls from the
nova-compute service to the nova-scheduler service via
the superconductor configuration in devstack, it becomes
increasingly important that we actually test the
affinity/anti-affinity server group scenarios.

Change-Id: Id3f4b1a8c6bb7c6f747ba28e7a3657d1e833595f