Do not run test_create_server_with_scheduler_hint_group twice
test_create_server_with_scheduler_hint_group will be run twice
in ServersTestJSON and ServersTestBootFromVolume with exactly
same behaviour, because it doesn't use the server created in
resource_setup.
This is to move test_create_server_with_scheduler_hint_group to
test_server_group.py, to let it be run only once, and to put it
together with server group testcases is reasonable.

test_create_server_with_scheduler_hint_group is not used in
interop, so we can safely move it.

Change-Id: Ie8c9b542e62965615e96c06009a7e6554f07fb06