Do not create server in test_list_servers_filter_by_exist_host
In test_list_servers_filter_by_exist_host, a server is created
using admin user, but in fact, we need not create that server,
instead, we can use the server created in resource_setup, because
admin user can list servers created by non-admin user.

This is to use the server created in resource_setup, both to make
the testcase faster and to make the code cleaner.

Change-Id: I89d0cdd31f26748f1364cfa57a80984566b3f850