Merge volume message list and show testcases
test_list_messages and test_show_message create a volume
type and a volume respectively which is not necessary,
in other words, they can share a volume type and volume
and message, so we can spare one create_volume_type and
one create_volume.

We can either create a volume type and a volume in
resource_setup for sharing usage, or to merge test_list_messages
and test_show_message, but since UserMessagesTest uses
min_microversion which can only have effect after setUp,
that means we can't call list_messages in resource_setup,
so this is to merge list and show into one testcase.

Change-Id: I68fa98697dfb74ec16e4bdb970bad6737817e70b