Stop using resource_cleanup in identity_v3 tests
Replace resource_cleanup with addClassResourceCleanup in identity
v3 tests.

Besides, in test_update_endpoint, service2 should be created before
endpoint_for_update is created, because Service2 must be deleted after
endpoint_for_update is deleted, otherwise we will get a 404 error
when deleting endpoint_for_update if endpoint's service is deleted.

And in test_domain_configuration, group created in resource_setup
is not used anywhere, so to remove it.

Change-Id: I00f5308f953bac191697fb7312b5f56093e80039