Disable testing of the v2.0 identity API
The v2.0 identity API is being removed in the Queens release, but in
order to do so we need to exclude/remove some v2.0 tempests tests
from being run. Since tempest is branchless we should keep the tests
around until Queens is EOL and then we can go through and clean up
all the v2.0 cruft from tempest.

Change-Id: I5d6bd5e1bfe2d5c999ced9ff92de846e50cc2c12