test_mtu_sized_frames: sanity check connectivity first
When the test case fails, it's often hard to distinguish between the
actual case of MTU-sized frames not allowed through without
fragmentation, or a general connectivity issue that affects the server.
To simplify debugging failures, the patch makes tempest first validate
that connectivity works at all for the instance before switching to
validating a more specific scenario.

Change-Id: I8f254ff79806533bf8a297e1bc07ed2216d8940c