Fix senario test: test_swift_acl_anonymous_download
In Zuul and Jenkins gate jobs like [1][2], the senario test:
test_swift_acl_anonymous_download is always failed, which is in
file test_object_storage_basic_ops.py. The reason is that the function
'change_container_acl' in tempest.senario.manager.ObjectStorageSenarioTest
is implemented incorrectly. This patch fixes the problem.

[1] legacy-tempest-dsvm-neutron-scenario-multinode-lvm-multibackend
[2] gate-tempest-dsvm-neutron-scenario-multinode-lvm-multibackend-ubuntu-xenial-nv

Closes-bug: #1722164
Change-Id: I4b3b6f4d7c5fd5e1e968b0144ecc2a7b3eae206b