Remove wrapper methods from object_client
There are few methods in object_clients which are
wrapper on PUT API with different combination of
headers.

As object_client is going to be declared as lib interface
it is better to remove these wrapper.

NOTE: Those methods are not being used outside of Tempest.

Partially implements blueprint consistent-service-method-names

Change-Id: Ia70ef3913c1650fad7b8f6b9019378b074c88155