Setup networking for live migration tests
It turns out the live migration tests have always
been creating an instance with no networking, so
we haven't been testing any of the network setup
and port binding update logic that happens during
a live migration, which is bad.

This sets up networking so the instance will get
a port when it's created and then the port binding
should get updated once the live migration is complete.

Change-Id: I41cfae9a7912aa8e2f0010a7ec2fdbd45ebf5e9d