Move and update test_resize_volume_backed_server_confirm
There is a bug in nova where the libvirt driver incorrectly removes
the local guest files for the hypervisor from shared storage because
it does not realize the instance is volume-backed. Nova runs an NFS
CI job in the experimental queue but it's not failing on resize
tests because they aren't volume-backed, so to recreate the failure
and make sure we don't regress the fix, this test is added. Nova
also runs a Ceph job which applies here too.

This moves the existing scenario test to be a compute API test. The
test ID is maintained in case people are blacklisting it, but the
slow tag is dropped because it's not a particularly slow test and
if the slow tag is applied, it won't actually be run in the jobs
that we care about testing this, e.g. NFS and Ceph jobs.

The additional wrinkle of getting the console log after the resize
is what shows the failure when the bug is not fixed, so that's added
here.

Depends-On: I29fac80d08baf64bf69e54cf673e55123174de2a

Change-Id: Id7de5186b2ea0ff7af86d9950c69203914498d88
Related-Bug: #1728603