Fix AssertionError in test_rand_password_with_len_2
This commit fixes the AssertionError in test_rand_password_with_len_2.
We were thinking an error could be happened but it's really rare.
However, it was actually happened[1]. To make the test more stable,
we shouldn't check its value but only length. It should be enough
because we already checked values in the other test cases.

[1] http://paste.openstack.org/show/625330/

Change-Id: I9a11a56e5eee9d1b01b61c04ea4bc75037c4458e