Add a .htaccess for redirecting old paths
This commit adds an .htaccess file to the tempest docs to handle
redirecting old paths to the new location. Before the docs migration in
the past we had docs located at docs.openstack.org/developer/tempest and
now things are located in docs.openstack.org/tempest/latest. With this
file we should be properly redirecting old links to the new location.

Change-Id: Idec7b2f299ade9b7dcc678ea1daf8daa25f08c44