Fix directory tree appearance in document
This commit fixes the directory tree appearance in document. It looks
like that the original rst line blocks don't work in the current
renderer. As a result, the directory structure is a little difficult to
understand. So, this commit use a literal block instead of that as a
quick fix.

Change-Id: I897cec40cc1ee5ead50997e75c474e1f78ee763e