Refactor volume groups and group snapshots tests
According to api microversion history of Cinder[1], min microversion for
volume groups (create/delete/update/list/show) is 3.13, min microversion
for group snapshots(create/delete/list/show) and create group from group
or group snapshot is 3.14.

While in GroupsTest in Tempest, the min microversion for volume groups is
set to 3.14. This patch aims to refactor the original GroupsTest to correct
the min microversion for volume groups, and also spilt it into two parts,
one for groups tests and the other for group snapshot tests, which will make
the code more clear and readable.

[1] https://docs.openstack.org/cinder/latest/contributor/api_microversion_history.html

Change-Id: Ic98e8892c4a44850d7a7afac9be1aa59ab38bb8c