Move test_tokens test to static
Letting os_primary be the user in this test case allows the token to
authenticate, get and delete without having to create a test user. This
allows the test to work with pre-provisioned credentials. Also moves the
test to the non-admin directory because the test uses os_primary creds
(for principle of least privilege).

Partial-Bug: #1714277
Co-Authored-By: Tin Lam <tin@irrational.io>
Change-Id: I55345132e0f461b36b08d222680a7e11eb945116