Add addOnException of unshelve_server when using sharing server
This is to add addOnException of unshelve_server when the sharing
server is shelved, otherwise if the testcase fail unexpectedly
the sharing server can't reach ACTIVE status and has to be recreated
after the timeout period is expired.

We can't use addCleanup because the steps added by addCleanup will
be executed after tearDown.

Change-Id: I84330d7bf24e4d693a42fb604fd728ad99bfb5cf