Only use stestr and tempest run in tempest tox jobs
Now that we're starting to unify everything on a single test runner we
should only be using either tempest run or stestr (which tempest run
will use internally in the near future). This commit changes the
coverage job to use stestr and removes the ostestr job as part of that
effort.

Change-Id: I5972346b8dbef4a57c733f85ceeeeca425bf263e