Add release note about the ostestr tox job
The ostestr tox job was recently removed from the tempest tox.ini,
mostly because we're moving towards standardizing on just using stestr
for everything and dropping the dependency on os-testr. During the
review that removed the job, some concerns were raised that it might
break people. To address that this commit adds a release note explaining
the potential upgrade implications and how to work around the removal.

Change-Id: Ibc59e01af9f5a9c4d8c67fbde1f029decf602226