Revert "Update volume-status waiter for new cinder attach"
This reverts commit 030dd17f5e0862384dec4c0b201d66f97074c5fb.

This masks a behavior change in the compute API when attaching
a volume to a shelved offloaded instance, as indicated by the
need to change Tempest to wait for a volume to be 'reserved'
vs 'in-use' in this scenario with the new Cinder v3 attach
flow.

This is a problem that needs to be handled in Nova to keep
API compatibility with the old volume attachment flow. Nova
can either do that under the covers, or add a new
microversion for callers to opt into this new behavior, but
masking over it in Tempest is wrong.

The related nova change:

  Ifc01dbf98545104c998ab96f65ff8623a6db0f28

Change-Id: If8e1f42d853f366a399c9e454a80ba3bf8cd136e