Add volume-attached resize revert server test
We have a test to confirm a volume-backed server but there
are subtle differences in how nova handles volume-backed servers
vs attached non-root volumes during a resize, and we regressed
that functionality recently because we don't have CI coverage
for it.

Trove CI jobs have a test for a non-root volume attached resize
confirm scenario but the resize revert scenario is not tested
in integration CI so a test is added for that here.

Depends-On: Ifc80d07d94311534fd9e7824ede9d09223a011c2

Change-Id: I9b8433bcbd45c2a0c5b38e17e3a86d3b0ac68d08
Related-Bug: #1737599