Do not use class variable for clients in ContainerSyncTest
Running tempest.api.object_storage.test_container_sync_middleware.
ContainerSyncMiddlewareTest.test_container_synchronization along with
all other swift testcases will fail with Unauthorized error, but it can
pass when being executed alone.

This is because clients in parent class ContainerSyncTest is a class
variable, which will cause trouble when using dynamic credentials. So
this is to change clients to an instance variable.

Change-Id: I10f04fadae95932258a5df58ef68d80f81075b79
Closes-Bug: #1740075