Do not test create_interface_by_fixed_ips if no isolated network
By default only project that is admin or network owner or project
with role advsvc is authorised to create interfaces with fixed-ip,
so if we don't create network for each project, do not test
_test_create_interface_by_fixed_ips.

This is to split test_create_list_show_delete_interfaces to two testcases,
so we can get to know that whether we tested interface by fixed_ips or not.
And test_create_list_show_delete_interfaces is not used in interop so
we can safely change it.

/etc/neutron/policy.json:
"create_port:fixed_ips": "rule:context_is_advsvc or rule:admin_or_network_owner"

description for context_is_advsvc:
https://github.com/openstack/neutron/commit/d4f00659eb18c6a5df2ce6d35202b0a4a9409f70

Change-Id: Ia4b6083eb7e5d8a69aff949d301ad5a118180bd9