Do not select inactive l3 agent in L3AgentSchedulerTestJSON
There may be some inactive l3 agents in the system, because the
system won't delete l3 agent records that ever registered,
even if they are no longer valid.

So this is to choose only active l3 agent in resource_setup,
otherwize the testcase will fail unexpectedly because we happened
to choose an inactive l3 agent.

Change-Id: Ie14cafc883c92e791ce866d548e166702bed3d0b