Revert "Use a non admin privileges for retyping a volume"
This reverts commit f5869b4557855af9450501c65e03bd16fd269036.

When retyping a volume, it is possible that Cinder falls back to
migrating a volume in order to complete the retype which is admin-only
operation.

This will result in a failure because Cinder will attempt to contact
Nova using the credentials of the user which are not elevated, with
Nova refusing to do the volume swap.

Therefore, in order to effectively test this, we should stick to using
the admin API as a retype can require a migration which is an
admin-only operation.

Related-Bug: #1740544
Change-Id: I3e4853146bb6e2a62205ffe690da081229215f54