Refactor config drive mounting
There are a few places where we mount the config drive which
is not done in a consistent manner which results in failures
for the cases where the config drive is VFAT, which makes the
device label report as CONFIG-2 and not config-2 (uppercase).

This refactors the mounting and unmounting codebase to the
same block and reuses it to ensure a consistent behaviour across
tests, also letting the tags checks pass when using VFAT for
config drives.

Change-Id: I529cff87c848bc87d63d2ec407ed18cdd631ecef