Connect to default ports if none are specified
If the VNC is running at a default port and the address returned
is an address which does not contain a port, then the VNC tests
fail to connect because they attempt to call a socket connect
without an IP listed.

In addition, this makes sure that the Host header does not include
a port if we are hitting the default port so things are clean.

Change-Id: Ie5939a4c5789efb759f5a2372435be7f9af33e91