Decouple AttachVolumeShelveTestJSON from AttachVolumeTestJSON
Since AttachVolumeShelveTestJSON was extending AttachVolumeTestJSON
to re-use common methods and setup, it was also running the tests
from the base class which have nothing to do with shelve/unshelve
scenarios, and adding around 3 minutes to each Tempest run.

This decouples the classes that run the actual tests by moving
the common code and setup into a common base class which does not
have any tests.

The max_microversion restriction is removed from AttachVolumeTestJSON
because there is nothing microversion-specific about the tests in
that class and it was added with AttachVolumeShelveTestJSON when it
extended AttachVolumeTestJSON.

Change-Id: I8cb9cbb162361cbf4adbea11b8ffaa7f7c45485f