Wait for server to be active in test_device_tagging
In test_device_tagging, we should wait for server to be
active before getting ssh_client. This testcase doesn't
fail in ci because it uses validation.connect_method = floating,
which will wait for server to be active if validatable=True
is passed in.

Besides, before get_server_ip we should call show_server,
otherwise 'addresses' is not present in the result of create_server
and so we will get:
File "/tempest/api/compute/base.py", line 387, in get_server_ip
    addresses = server['addresses'][CONF.validation.network_for_ssh]
    KeyError: 'addresses'

Change-Id: I70410c47394310795064cd9fde0b9b47ddd00982