Handle creating an image from a server with microversion > 2.35
The compute 2.35 API deprecates the compute image proxy APIs
for showing/listing/deleting images.

The compute 2.45 API removes the 'location' header from the createImage
server action response and in that same version, adds a response body
with a single "image_id" key.

This change updates the BaseV2ComputeTest.create_image_from_server
helper method to be aware of the microversion that the test running
this code is using, and adjust the client to use appropriately.

As a result, the compute images client needs to be aware of the
changed createImage response schema so that is added for compute
API version 2.45.

Change-Id: I5551af0064f9cca594ae533379d0b0ae14444f88