Do not skip entire testcase if floating_ip is not enabled
TestMinimumBasicScenario and TestServerBasicOps are useful
and basic scenario testcases, in which floating_ip is not an
essential part and just a connecting method, so it's not
reasonable to skip entire testcase if floating_ip is not enabled.
This is to skip the creating of floating_ip when floating_ip
is not enabled, and use fixed_ip to ssh the server instead,
so we can still keep the most part of the test steps.

Change-Id: I3e99304462950c194a4d73237cfc594631ca3868