CLI tests: Identity v3 (project name, api version)
Switch to --os-project-name instead of --os-tenant-name, and pass
also --os-identity-api-version.
All the clients which are wrapped by some method of the CLIClient class
and which use cmd_with_auth support the newly introduced flags at least
since Liberty.
Only exception: the keystone client, which was removed in Newton, so
it is not supported anyway (and it has not been fixed).

Closes-Bug: #1721553
Change-Id: I6d34e76b4089be024093a75f9e467d273d80f1c2