Restrict test_list_endpoints_for_token to tempest
The test_list_endpoints_for_token test now checks for services in
service_available, however that may include services from plugins.
We don't have a strong contract on the meaning of service_available
so this is not correct as it will look in the catalog for things
that may not even be OpenStack services.

Fixing this by restricting by services in Tempest (including keystone).

Change-Id: I32f4f056ea49fe04bfc309abe2cc9b9234b92847
Closes-bug: #1743996