Make tempest-full-py3 voting
The job is ready to replace the legacy one in the integration
gate py3 template. Before that can happen make it voting and add
some extra documentation to it.

The parent does not need to be tempest-full since that does not
add much useful as opposed to devstack-tempest and it's actually
better for the two jobs to be independent.

Change-Id: I4a8fa1759665038788e1b86fd3a5030fd6c68573