Add the nova-multiattach job to the check/gate queues
The nova-multiattach job is defined in the nova repo
but we also need to run it against Tempest changes.

For now the job just runs the tempest compute API tests
since that's the only place in Tempest that has multiattach
testing right now and it will save some time in the CI run.

Change-Id: Idfb68b5ddd7aca8642016e392fec6efd169c1ea2
Depends-On: I51adbbdf13711e463b4d25c2ffd4a3123cd65675