Fixes test_device_tagging IndexError
If the metadata_service config option is set to True, the test
test_device_tagging will connect to the spawned instance through ssh
and curl the metadata service and checks the response for the assigned
device tags.

However, not all devices are tagged, causing an IndexError. This patch
solves this issue.

Libvirt only exposes tagged devices, but other drivers (Hyper-V) can
expose other devices as well.

Change-Id: I7570cd81eec09109ad5c74315a7876f53c032240
Closes-Bug: #1747872