Added storyboard integration to tempest.lib decorators
A new parameter is introduced to related_bug/skip_because
decorators called `bug_type` which takes in either 'launchpad'
(default for backward compatibility) or 'storyboard'. The
appropriate link for each tracking page is generated based off
this value and the 'bug' value.

This is useful for projects like Monasca Tempest plugin which
tracks issues on Storyboard. Also,  Storyboard is the new community
preferred place to track stories and bugs [0]. So Tempest
should provide projects with the ability to track bugs using it.

Note that Storyboard does **not** only track stories. It tracks
bugs too: "It all begins with a story. A story is a bug report or
proposed feature. Stories are then further split into tasks,
which affect a given project and branch."

Unit tests and releasenotes are included.

[0] https://wiki.openstack.org/wiki/StoryBoard
[1] https://storyboard.openstack.org/#!/page/about

Co-Authored-By: Felipe Monteiro <felipe.monteiro@att.com>
Change-Id: Ic34208cfe997ceacdafd1ce122691b58a9778e78