Add simple compute API test to create server with utf8 name
Copied from test_create_image_specify_multibyte_character_image_name
this is a similar test which creates a server with a multi-byte
unicode name.

This is needed to verify what appears to be a regression in nova
in the queens release, and to make sure we don't regress again in
the future.

Depends-On: I7afce618deca3baaa96605a6b48c1c2ef4a5f2a5

Change-Id: If5a76dc3b03e5f7ceb4c03b3555318170e2ce69f
Related-Bug: #1730756