Deprecate all-plugin
The all-plugin tox env uses sitepackages so that plugins
installed outsite of Tempest virtual environment can be discovered.
After the implementation during the Queens release cycle of the
goal of moving Tempest plugins in dedicated git repos, this
environment should not be used anymore. "all" should be used
instead with the appropriate regex filtering.

The deprecation is introduced only in the form of a comment in
tox.ini and a WARNING echoed before the test run.

Patrole needs an environment with site-packages on, so create a
new one identical to the old all-plugin, but with a less confusing
name, for patrole to use while avoiding deprecation warnings.

Change-Id: Ib97fb3d0ce64b6faac6a190afc0006b624ce0a62